# Enhance search performance