@mourner 2025-05-10
Also wondering: is this increase for smaller queries a fluke due or is there an overhead?

> before: 1000 searches 0.01%: 11.675ms
> after: 1000 searches 0.01%: 15.793ms