@muendlein 2025-05-11
> Also wondering: is this increase for smaller queries a fluke due or is there an overhead?
> 
> > before: 1000 searches 0.01%: 11.675ms
> > after: 1000 searches 0.01%: 15.793ms

I would say both is possible here. The current benchmark is not sophisticated enough to track this down. Overall the time for small queries seems to scatters signicantly with every run.